refactor: move ecstore pool-space builder
