+++ crates/ecstore/src/store/peer.rs
@@ -239,47 +239,3 @@ pub async fn get_disk_infos(disks: &[Option<DiskStore>]) -> Vec<Option<DiskInfo>
 
     res
 }
-
-pub async fn has_space_for(dis: &[Option<DiskInfo>], size: i64) -> Result<bool> {
-    let size = { if size < 0 { DISK_ASSUME_UNKNOWN_SIZE } else { size as u64 * 2 } };
-
-    let mut available = 0;
-    let mut total = 0;
-    let mut disks_num = 0;
-
-    for disk in dis.iter().flatten() {
-        disks_num += 1;
-        total += disk.total;
-        available += disk.total - disk.used;
-    }
-
-    if disks_num < dis.len() / 2 || disks_num == 0 {
-        return Err(Error::other(format!(
-            "not enough online disks to calculate the available space,need {}, found {}",
-            (dis.len() / 2) + 1,
-            disks_num,
-        )));
-    }
-
-    let per_disk = size / disks_num as u64;
-
-    for disk in dis.iter().flatten() {
-        if !is_erasure_sd().await && disk.free_inodes < DISK_MIN_INODES && disk.used_inodes > 0 {
-            return Ok(false);
-        }
-
-        if disk.free <= per_disk {
-            return Ok(false);
-        }
-    }
-
-    if available < size {
-        return Ok(false);
-    }
-
-    available -= size;
-
-    let want = total as f64 * (1.0 - DISK_FILL_FRACTION);
-
-    Ok(available > want as u64)
-}
